requirements.txt
PySimpleGUI
requests
python-dotenv
pillow